+++ Controller/Payment/UpdateQuotePaymentMethod.php
@@ -21,7 +21,7 @@ class UpdateQuotePaymentMethod extends Action
     /**
      * @var JsonFactory
      */
-//    private $jsonResultFactory;
+    private $jsonResultFactory;
     
     private $readerWriter;
 
@@ -35,13 +35,13 @@ class UpdateQuotePaymentMethod extends Action
     public function __construct(
         Context $context,
         ModuleConfig $moduleConfig,
-//        JsonFactory $jsonResultFactory,
+        JsonFactory $jsonResultFactory,
         \Nuvei\Checkout\Model\ReaderWriter $readerWriter
     ) {
         parent::__construct($context);
 
         $this->moduleConfig         = $moduleConfig;
-//        $this->jsonResultFactory    = $jsonResultFactory;
+        $this->jsonResultFactory    = $jsonResultFactory;
         $this->readerWriter         = $readerWriter;
     }
 
@@ -50,8 +50,7 @@ class UpdateQuotePaymentMethod extends Action
      */
     public function execute()
     {
-        $result = $this
-            ->jsonResultFactory
+        $result = $this->jsonResultFactory
             ->create()
             ->setHttpResponseCode(\Magento\Framework\Webapi\Response::HTTP_OK);